requirement.txt
cffi==1.15.1
docstring-parser==0.14.1
executing==1.2.0
msgpack==1.0.5
py-algorand-sdk==2.4.0
pycparser==2.21
pycryptodomex==3.18.0
PyNaCl==1.5.0
pyteal==0.25.0
semantic-version==2.10.0
tabulate==0.9.0
